fixed navbar and about page route
