+++ src/app/(nav)/about/Sections/Hero.jsx
