+++ src/app/(nav)/about/Sections/HowItWorks.jsx
