+++ src/app/(nav)/about/Sections/StorySection.jsx
